moved home route to centralized event logic
